chore: More colors to own files + small color tweaks
